Add files via upload
# alignaudio() MATLAB Function.

This script uses MATLAB's alignsignals() function to align two audio files. The delay time calculated by this function is used to displace the the leading signal to bring it into alignment with the lagging signal. Once the alignment has been carried out, the shortest of the two signals undergoes zero-padding at the end to ensure that the two output signals are the same sample length. The alignsignals() function calculates the autocorrelation of the two signals and uses the peak correlation value to determine the displacement value for the alignment.

The script only aligns single channels, therefore if more than one channel is present only the left channel is used.

## Requirements and Info

- Sample rates of audio files must be the same.
- Only mono. Multi-channel audio files will be read, however only the left channel will be processed for alignment.